chore(.kodiak.toml): add .kodiak.toml configuration file for Kodiak CI
The .kodiak.toml file is added to the repository to configure the
behavior of the Kodiak CI tool. This file specifies the version of the
configuration, sets up automatic merging of pull requests labeled with
"dependencies" or "automerge", and defines the merge method as "rebase".
It also configures automatic merging of PRs opened by the "dependabot"
user that are "minor" or "patch" version upgrades, while ignoring
"major" version upgrades. Additionally, it specifies that the
"dependabot" user should be ignored when updating and closing stale
dependency upgrades.