fix(ci.yml): update DockerHub credentials in the CI workflow
The DockerHub credentials in the CI workflow have been updated to use
the correct secrets. The `DOCKER_USER` secret has been changed to
`DOCKERHUB_USER` and the `DOCKERHUB_PASSWORD` secret has been changed to
`DOCKER_PASS`. This ensures that the correct credentials are used for
logging into DockerHub and caching Docker layers during the CI process.